+++ www/src/pages/ecosystem.js
@@ -1,13 +1,12 @@
 import React, { Component } from "react"
 import { graphql } from "gatsby"
-import { Helmet } from "react-helmet"
 
 import Layout from "../components/layout/layout-with-heading"
 import EcosystemBoard from "../components/ecosystem/ecosystem-board"
+import PageMetadata from "../components/page-metadata"
 import FooterLinks from "../components/shared/footer-links"
 
 import { EcosystemIcon } from "../assets/icons"
-import { PluginsIcon, StartersIcon } from "../assets/icons/ecosystem-icons"
 
 class EcosystemPage extends Component {
   render() {
@@ -43,22 +42,15 @@ class EcosystemPage extends Component {
     const plugins = pluginsData
 
     const pageTitle = `Ecosystem`
-    const boardIcons = { plugins: PluginsIcon, starters: StartersIcon }
 
     return (
       <Layout
         location={location}
         pageTitle={pageTitle}
-        pageIcon={EcosystemIcon}
+        pageIcon={<EcosystemIcon />}
       >
-        <Helmet>
-          <title>Ecosystem</title>
-        </Helmet>
-        <EcosystemBoard
-          icons={boardIcons}
-          starters={starters}
-          plugins={plugins}
-        />
+        <PageMetadata title="Ecosystem" />
+        <EcosystemBoard starters={starters} plugins={plugins} />
         <FooterLinks />
       </Layout>
     )