+++ www/src/templates/tags.js
@@ -1,7 +1,8 @@
 /** @jsx jsx */
 import { jsx } from "theme-ui"
 import React from "react"
-import { Helmet } from "react-helmet"
+import styled from "@emotion/styled"
+import { mediaQueries } from "gatsby-design-tokens/dist/theme-gatsbyjs-org"
 import { graphql } from "gatsby"
 import { TiTags as TagsIcon, TiArrowRight } from "react-icons/ti"
 
@@ -9,6 +10,7 @@ import BlogPostPreviewItem from "../components/blog-post-preview-item"
 import Button from "../components/button"
 import Container from "../components/container"
 import FooterLinks from "../components/shared/footer-links"
+import PageMetadata from "../components/page-metadata"
 import { TAGS_AND_DOCS } from "../data/tags-docs"
 
 // Select first tag with whitespace instead of hyphens for
@@ -23,6 +25,16 @@ const preferSpacedTag = tags => {
   return tags[0]
 }
 
+const ButtonWrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+
+  ${mediaQueries.xs} {
+    flex-direction: row;
+  }
+`
+
 const Tags = ({ pageContext, data }) => {
   const { tags } = pageContext
   const { nodes, totalCount } = data.allMdx
@@ -33,36 +45,35 @@ const Tags = ({ pageContext, data }) => {
 
   return (
     <Container>
-      <Helmet>
-        <title>{`${preferSpacedTag(tags)} Tag`}</title>
-        <meta
-          name="description"
-          content={`Case studies, tutorials, and other posts about Gatsby related to ${preferSpacedTag(
-            tags
-          )}`}
-        />
-      </Helmet>
+      <PageMetadata
+        title={`${preferSpacedTag(tags)} Tag`}
+        description={`Case studies, tutorials, and other posts about Gatsby related to ${preferSpacedTag(
+          tags
+        )}`}
+      />
       <h1>{tagHeader}</h1>
-      <Button
-        variant="small"
-        key="blog-post-view-all-tags-button"
-        to="/blog/tags"
-      >
-        View all tags <TagsIcon />
-      </Button>
-      {doc ? (
-        <React.Fragment>
-          <span css={{ margin: 5 }} />
-          <Button
-            variant="small"
-            secondary
-            key={`view-tag-docs-button`}
-            to={doc}
-          >
-            Read the documentation <TiArrowRight />
-          </Button>
-        </React.Fragment>
-      ) : null}
+      <ButtonWrapper>
+        <Button
+          variant="small"
+          key="blog-post-view-all-tags-button"
+          to="/blog/tags"
+        >
+          View all tags <TagsIcon />
+        </Button>
+        {doc ? (
+          <React.Fragment>
+            <span css={{ margin: 5 }} />
+            <Button
+              variant="small"
+              secondary
+              key={`view-tag-docs-button`}
+              to={doc}
+            >
+              Read the documentation <TiArrowRight />
+            </Button>
+          </React.Fragment>
+        ) : null}
+      </ButtonWrapper>
       {nodes.map(node => (
         <BlogPostPreviewItem
           post={node}
@@ -84,8 +95,7 @@ export const pageQuery = graphql`
       sort: { fields: [frontmatter___date, fields___slug], order: DESC }
       filter: {
         frontmatter: { tags: { in: $tags } }
-        fileAbsolutePath: { regex: "/docs.blog/" }
-        fields: { released: { eq: true } }
+        fields: { section: { eq: "blog" }, released: { eq: true } }
       }
     ) {
       totalCount