+++ www/src/utils/node/starters.js
@@ -1,15 +1,13 @@
 const _ = require(`lodash`)
 const Promise = require(`bluebird`)
-const path = require(`path`)
-const fs = require(`fs-extra`)
-const { slash } = require(`gatsby-core-utils`)
 const getpkgjson = require(`get-package-json-from-github`)
 const parseGHUrl = require(`parse-github-url`)
 const { GraphQLClient } = require(`@jamo/graphql-request`)
-const yaml = require(`js-yaml`)
-const ecosystemFeaturedItems = yaml.load(
-  fs.readFileSync(`./src/data/ecosystem/featured-items.yaml`)
+const { loadYaml } = require(`../load-yaml`)
+const { starters: featuredStarters } = loadYaml(
+  `src/data/ecosystem/featured-items.yaml`
 )
+const { getTemplate } = require(`../get-template`)
 
 if (
   process.env.gatsby_executing_command === `build` &&
@@ -29,12 +27,47 @@ const githubApiClient = process.env.GITHUB_API_TOKEN
     })
   : null
 
+exports.sourceNodes = ({ actions: { createTypes } }) => {
+  createTypes(/* GraphQL */ `
+    type StartersYaml implements Node {
+      url: String!
+      repo: String!
+      description: String
+      tags: [String!]
+      features: [String!]
+      childScreenshot: Screenshot # added by gatsby-transformer-screenshot
+      fields: StartersYamlFields!
+    }
+
+    type StartersYamlFields @dontInfer {
+      featured: Boolean
+      hasScreenshot: Boolean
+      starterShowcase: StartersYamlFieldsStarterShowcase
+    }
+
+    type StartersYamlFieldsStarterShowcase @dontInfer {
+      slug: String!
+      stub: String
+      name: String
+      description: String
+      stars: Int
+      lastUpdated: String
+      owner: String
+      githubFullName: String
+      gatsbyMajorVersion: [[String]]
+      allDependencies: [[String]]
+      gatsbyDependencies: [[String]]
+      miscDependencies: [[String]]
+    }
+  `)
+}
+
 exports.createPages = async ({ graphql, actions, reporter }) => {
   const { createPage } = actions
 
-  const starterTemplate = path.resolve(`src/templates/template-starter-page.js`)
+  const starterTemplate = getTemplate(`template-starter-page`)
 
-  const { data, errors } = await graphql(`
+  const { data, errors } = await graphql(/* GraphQL */ `
     query {
       allStartersYaml {
         nodes {
@@ -68,10 +101,10 @@ exports.createPages = async ({ graphql, actions, reporter }) => {
     }
   })
 
-  starters.forEach((node, index) => {
+  starters.forEach(node => {
     createPage({
       path: `/starters${node.fields.starterShowcase.slug}`,
-      component: slash(starterTemplate),
+      component: starterTemplate,
       context: {
         slug: node.fields.starterShowcase.slug,
       },
@@ -79,10 +112,10 @@ exports.createPages = async ({ graphql, actions, reporter }) => {
   })
 }
 
-const fetchGithubData = async ({ owner, repo, reporter }, retry = 0) => {
-  return githubApiClient
+const fetchGithubData = async ({ owner, repo, reporter }, retry = 0) =>
+  githubApiClient
     .request(
-      `
+      /* GraphQL */ `
     query {
       repository(owner:"${owner}", name:"${repo}") {
         name
@@ -106,7 +139,7 @@ const fetchGithubData = async ({ owner, repo, reporter }, retry = 0) => {
       }
 
       const githubUrl = `https://github.com/${owner}/${repo}`
-      const { url } = await fetch(githubUrl, { method: "HEAD" })
+      const { url } = await fetch(githubUrl, { method: `HEAD` })
       const { owner: newOwner, name: newRepo } = parseGHUrl(url)
 
       reporter.warn(
@@ -118,10 +151,9 @@ const fetchGithubData = async ({ owner, repo, reporter }, retry = 0) => {
         retry + 1
       )
     })
-}
 
 exports.onCreateNode = ({ node, actions, getNode, reporter }) => {
-  const { createNodeField } = actions
+  const { createNodeField, deleteNode } = actions
   if (node.internal.type === `StartersYaml` && node.repo) {
     // To develop on the starter showcase, you'll need a GitHub
     // personal access token. Check the `www` README for details.
@@ -129,7 +161,7 @@ exports.onCreateNode = ({ node, actions, getNode, reporter }) => {
     const { owner, name: repoStub } = parseGHUrl(node.repo)
 
     // mark if it's a featured starter
-    if (ecosystemFeaturedItems.starters.includes(`/${owner}/${repoStub}/`)) {
+    if (featuredStarters.includes(`/${owner}/${repoStub}/`)) {
       createNodeField({ node, name: `featured`, value: true })
     }
 
@@ -156,7 +188,7 @@ exports.onCreateNode = ({ node, actions, getNode, reporter }) => {
     createNodeField({ node, name: `hasScreenshot`, value: !!screenshotNode })
 
     if (!process.env.GITHUB_API_TOKEN) {
-      return createNodeField({
+      createNodeField({
         node,
         name: `starterShowcase`,
         value: {
@@ -164,7 +196,7 @@ exports.onCreateNode = ({ node, actions, getNode, reporter }) => {
         },
       })
     } else {
-      return Promise.all([
+      Promise.all([
         getpkgjson(node.repo),
         fetchGithubData({ owner, repo: repoStub, reporter }),
       ])
@@ -184,9 +216,9 @@ exports.onCreateNode = ({ node, actions, getNode, reporter }) => {
           )
 
           const gatsbyMajorVersion = allDependencies
-            .filter(([key, _]) => key === `gatsby`)
+            .filter(([key]) => key === `gatsby`)
             .map(version => {
-              let [gatsby, versionNum] = version
+              const [gatsby, versionNum] = version
               if (versionNum === `latest` || versionNum === `next`) {
                 return [gatsby, `2`]
               }
@@ -208,11 +240,11 @@ exports.onCreateNode = ({ node, actions, getNode, reporter }) => {
             allDependencies,
             gatsbyDependencies: allDependencies
               .filter(
-                ([key, _]) => ![`gatsby-cli`, `gatsby-link`].includes(key) // remove stuff everyone has
+                ([key]) => ![`gatsby-cli`, `gatsby-link`].includes(key) // remove stuff everyone has
               )
-              .filter(([key, _]) => key.includes(`gatsby`)),
+              .filter(([key]) => key.includes(`gatsby`)),
             miscDependencies: allDependencies.filter(
-              ([key, _]) => !key.includes(`gatsby`)
+              ([key]) => !key.includes(`gatsby`)
             ),
           }
           createNodeField({
@@ -222,10 +254,11 @@ exports.onCreateNode = ({ node, actions, getNode, reporter }) => {
           })
         })
         .catch(err => {
-          reporter.panicOnBuild(
+          reporter.warn(
             `Error getting repo data for starter "${repoStub}":\n
             ${err.message}`
           )
+          deleteNode(node)
         })
     }
   }