+++ www/src/views/showcase/index.js
@@ -1,7 +1,8 @@
 import React, { Component } from "react"
-import { Helmet } from "react-helmet"
 import qs from "qs"
 import { navigate } from "gatsby"
+
+import PageMetadata from "../../components/page-metadata"
 import scrollToAnchor from "../../utils/scroll-to-anchor"
 import FeaturedSites from "./featured-sites"
 import FilteredShowcase from "./filtered-showcase"
@@ -49,13 +50,10 @@ class ShowcaseView extends Component {
 
     return (
       <>
-        <Helmet>
-          <title>Showcase</title>
-          <meta
-            name="description"
-            content="Gallery of sites using Gatsby across the web, find inspiration or inspect the code of popular projects."
-          />
-        </Helmet>
+        <PageMetadata
+          title="Showcase"
+          description="Gallery of sites using Gatsby across the web, find inspiration or inspect the code of popular projects."
+        />
         <FeaturedSites
           setFilters={this.setFilters}
           featured={data.featured.nodes}