+++ packages/gatsby-remark-responsive-iframe/CHANGELOG.md
@@ -3,6 +3,40 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.4.7](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-responsive-iframe@2.4.6...gatsby-remark-responsive-iframe@2.4.7) (2020-06-24)
+
+**Note:** Version bump only for package gatsby-remark-responsive-iframe
+
+## [2.4.6](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-responsive-iframe@2.4.5...gatsby-remark-responsive-iframe@2.4.6) (2020-06-22)
+
+**Note:** Version bump only for package gatsby-remark-responsive-iframe
+
+## [2.4.5](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-responsive-iframe@2.4.4...gatsby-remark-responsive-iframe@2.4.5) (2020-06-09)
+
+**Note:** Version bump only for package gatsby-remark-responsive-iframe
+
+## [2.4.4](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-responsive-iframe@2.4.3...gatsby-remark-responsive-iframe@2.4.4) (2020-06-02)
+
+**Note:** Version bump only for package gatsby-remark-responsive-iframe
+
+## [2.4.3](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-responsive-iframe@2.4.2...gatsby-remark-responsive-iframe@2.4.3) (2020-05-20)
+
+**Note:** Version bump only for package gatsby-remark-responsive-iframe
+
+## [2.4.2](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-responsive-iframe@2.4.1...gatsby-remark-responsive-iframe@2.4.2) (2020-05-05)
+
+**Note:** Version bump only for package gatsby-remark-responsive-iframe
+
+## [2.4.1](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-responsive-iframe@2.4.0...gatsby-remark-responsive-iframe@2.4.1) (2020-04-28)
+
+### Bug Fixes
+
+- update packages ([#23525](https://github.com/gatsbyjs/gatsby/issues/23525)) ([e65dd1e](https://github.com/gatsbyjs/gatsby/commit/e65dd1e))
+
+# [2.4.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-responsive-iframe@2.3.4...gatsby-remark-responsive-iframe@2.4.0) (2020-04-27)
+
+**Note:** Version bump only for package gatsby-remark-responsive-iframe
+
 ## [2.3.4](https://github.com/gatsbyjs/gatsby/compare/gatsby-remark-responsive-iframe@2.3.3...gatsby-remark-responsive-iframe@2.3.4) (2020-04-24)
 
 **Note:** Version bump only for package gatsby-remark-responsive-iframe